requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
google-cloud-storage==3.3.0
python-multipart==0.0.20
requests==2.32.5
gunicorn==23.0.0
pydantic==2.11.7
pyjwt==2.10.1
passlib[bcrypt]==1.7.4